Provider name changed
