+++ src/scraper/sites/manga/manganelo/ManganatoManagerUtils.ts
