+++ src/scraper/sites/manga/manganelo/managers/ManganatoManager.ts
